[client] clipboard/wayland: make use of boilerplate from wm.c
